Just a Test

+++ OOP/src/sam/oopintro/car/burnerCar.java
@@ -2,18 +2,18 @@ package sam.oopintro.car;
 
 public class burnerCar  extends Car {
 
-    private int fuelCapacity;
+    private double fuelCapacity;
 
-    public burnerCar(String name, int fuelCapacity) {
+    public burnerCar(String name, double fuelCapacity) {
         super(name);
         this.fuelCapacity = fuelCapacity;
     }
 
-    public int getFuelCapacity() {
+    public double getFuelCapacity() {
         return fuelCapacity;
     }
 
-    public void setFuelCapacity(int fuelCapacity) {
+    public void setFuelCapacity(double fuelCapacity) {
         this.fuelCapacity = fuelCapacity;
     }
 }
